+++ backend-v2/src/main/java/pt/ipleiria/careline/repositories/PatientRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pt.ipleiria.careline.domain.entities.users.PatientEntity;
 
 @Repository
-public interface PatientRepository extends JpaRepository<PatientEntity, Integer> {
+public interface PatientRepository extends JpaRepository<PatientEntity, Long> {
 }
